requirements.txt
openai>=1.14.2
python-dotenv>=1.0.1
pandas>=2.1.4
tqdm>=4.66.2
judges==0.0.10
deepeval==2.9.1